Started the profile page design. Created form item components for the sign up components for code re-use.
