+++ react-ui/src/Components/form-items/FirstNameFormItem.tsx
@@ -1,0 +1,24 @@
+import { Form, Input } from "antd"
+
+function FirstNameFormItem({
+  ...props
+}: React.ComponentProps<typeof Form.Item>) {
+  return(
+    <Form.Item
+      {...props}
+      name='firstName'
+      rules={[
+        // eslint-disable-next-line no-useless-escape
+        { pattern: /^\S*$/, message: 'First name cannot contain spaces.' },
+        { whitespace: true, message: 'Please enter your first name.' },
+        { required: true, message: 'Please enter your first name.' }
+      ]}
+    >
+      <Input
+        placeholder='First Name'
+      />
+    </Form.Item>
+  );
+}
+
+export default FirstNameFormItem
