+++ react-ui/src/css/Profile.css
@@ -1,7 +1,3 @@
 .profile-page-avatar {
   font-size: 50px !important;
 }
-
-.profile-page-sider {
-  padding-top: 12%;
-}
